# Conflicts:
#	modules/arrow/package.json
#	modules/arrow/tsconfig.json
#	modules/deck-arrow-layers/src/layers/arrow-layer-input.ts
#	modules/deck-arrow-layers/src/layers/arrow-path-layer.ts
#	modules/deck-arrow-layers/src/layers/arrow-polygon-layer.ts
#	modules/deck-arrow-layers/src/layers/arrow-text-layer.ts
#	modules/deck-arrow-layers/test/layers/arrow-layers.spec.ts
#	yarn.lock

+++ modules/arrow/test/arrow/arrow-text-source-mapping.node.spec.ts
@@ -1,8 +1,8 @@
 // luma.gl
 // SPDX-License-Identifier: MIT
-// Copyright (c) vis.gl contributors
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
 
-import test from '@luma.gl/devtools-extensions/tape-test-utils';
+import test from 'test/utils/vitest-tape';
 import {makeArrowFixedSizeListVector, resolveArrowTextSourceVectors} from '@luma.gl/arrow';
 import * as arrow from 'apache-arrow';
 
@@ -108,9 +108,9 @@ function makeArrowTextSourceVectors() {
     textAnchors: arrow.vectorFromArray([0, 1], new arrow.Uint8()),
     alignmentBaselines: arrow.vectorFromArray([0, 2], new arrow.Uint8()),
     clipRects: makeArrowFixedSizeListVector(
-      new arrow.Int16(),
+      new arrow.Float32(),
       4,
-      new Int16Array([0, 0, 8, 8, 0, 0, 8, 8])
+      new Float32Array([0, 0, 8, 8, 0, 0, 8, 8])
     )
   };
 }
